Created C version with module system
This version is capable of processing Illumina FASTQ format, including
changes in CASAVA 1.8. It also includes a module system to allow
validation of sequences.

The new version no longer does naïve assembly; it is now capable of
correcting errors in the sequence using the provided quality
information. It also does a better job of determining the optimal match.
Finally, it is significantly faster than the previous version.