Changed PANDA_MAX_LEN handling
Stupidly, the maximum sequence length specified could exceed the
sequence index in the hashtable. The solution is to assume that the
output sequence will be no more than 2*PANDA_MAX_LEN and assume that
PANDA_MAX_LEN is the length of the input sequences. All the code has
been changed to match this.