Fix bug where masks were applied to the wrong part of the sequence
The CASAVA pipeline masks low quality bases from the end of the read by
giving them all the PHRED score 2 (aka the B-cliff). This is detected by
the algorithm and the scores are treated differently. Commit
8487c61018c259f739bfcf99642dc06ca954b682 was supposed to mask the cliffs
in a new way (that did not avoid editing the sequence). However, it
inverted the sense of all the checks that were treating all the high
quality part of the sequence as low quality and using terrible low
quality scores that should be ignored in the remaining part.